Update TeamSpeakMonitorM5StickC.ino
Changed order of includes so Display.h is loaded before TeamSpeakFunctions.h.
Added code to set the display brightness to 0 if there are no clients and back to 96 if there are.
Replaced delay(500) with pause(500) in line 301, the only place where delay() is used at the moment during the main loop.